Refactor code for ListOfDirectories DB access
